chore(Editorial): moved stories assets
