+++ src/stories/assets/Editorial/portrait.png
